@@ -14,12 +14,13 @@ export class Converter {
         username: user.username,
         email: user.email,
         picUrl: user.picUrl,
-        timestamp: user.time,
+        time: user.time,
         recipes: user.recipes,
         wishlist: user.wishlist,
         shoppingList: user.shoppingList,
         following: user.following,
-        numRecipes: user.numRecipes
+        numRecipes: user.numRecipes,
+        aboutme: user.aboutme
       };
     },
     fromFirestore: function (
@@ -38,7 +39,8 @@ export class Converter {
         data.recipes,
         data.wishlist,
         data.shoppingList,
-        data.numRecipes
+        data.numRecipes,
+        data.aboutme
       );
     },
   };
