@@ -1,9 +1,12 @@
+import {Component, OnInit} from '@angular/core';
 import {
-  AngularFirestore
+  AngularFirestore,
+  AngularFirestoreCollection,
 } from '@angular/fire/firestore';
-import { Component, OnInit } from '@angular/core';
-
-import { AngularFireAuth } from '@angular/fire/auth';
+import {User} from '../user'
+import {Converter} from '../converter'
+import {Recipe} from '../recipe'
+import {RecipeConverter} from '../recipe-converter'
 
 @Component({
   selector: 'app-home',
@@ -11,8 +14,28 @@ import { AngularFireAuth } from '@angular/fire/auth';
   styleUrls: ['./home.component.scss'],
 })
 export class HomeComponent implements OnInit {
+  recipeCollection: AngularFirestoreCollection<Recipe> = this.afs.collection('recipes');
+  userCollection: AngularFirestoreCollection<User> = this.afs.collection('users');
+  newUsers!: firebase.firestore.QueryDocumentSnapshot<User>[];
+  topRecipes!: firebase.firestore.QueryDocumentSnapshot<Recipe>[];
+  newRecipes!: firebase.firestore.QueryDocumentSnapshot<Recipe>[];
 
-  constructor(public fAuth: AngularFireAuth, private afs: AngularFirestore) {}
+  constructor(private afs: AngularFirestore) {
+    
+    this.userCollection.ref.orderBy("time", "desc").limit(4).withConverter(new Converter().userConverter).get().
+        then((user) => {
+          this.newUsers = user.docs;
+        });
+    //timestamp for now. In future - change to 'ratings'
+    this.recipeCollection.ref.orderBy("images", "asc").limit(4).withConverter(new RecipeConverter().recipeConverter).get().
+        then((recipe) => {
+          this.topRecipes = recipe.docs;
+        });
+    this.recipeCollection.ref.orderBy("timestamp", "desc").limit(4).withConverter(new RecipeConverter().recipeConverter).get().
+        then((recipe) => {
+          this.newRecipes = recipe.docs;
+        });
+  }
 
   ngOnInit() {}
 }
