@@ -1,65 +1,139 @@
-import { Component } from '@angular/core';
-import { FormControl, FormGroup, Validators } from '@angular/forms';
-import { AngularFirestore } from '@angular/fire/firestore';
-import { Router } from '@angular/router';
-import { formattedError } from '@angular/compiler';
+import {Component, Output, NgZone, EventEmitter} from '@angular/core';
+import {FormArray, FormBuilder, FormControl, FormGroup, Validators} from '@angular/forms';
+import {AngularFirestore} from '@angular/fire/firestore';
+import {Router} from '@angular/router';
+import {AngularFireStorage} from '@angular/fire/storage';
+import {DomSanitizer, SafeUrl} from '@angular/platform-browser';
+import {AngularFireAuth} from '@angular/fire/auth'
+import {Recipe} from '../recipe';
+import {RecipeConverter} from '../recipe-converter';
+
 
 @Component({
   selector: 'app-upload-recipe',
   templateUrl: './upload-recipe.component.html',
-  styleUrls: ['./upload-recipe.component.scss'],
+  styleUrls: ['./upload-recipe.component.scss']
 })
 
-export class UploadRecipeComponent {
-  constructor(private db: AngularFirestore,
-              private router: Router,) {
-  }
+export class UploadRecipeComponent{
+  error = '';
+  hide = true;
+  imgFile!: Blob;
+  pictureUrl = '';
+  progress = 0;
+  previewImgUrls: SafeUrl[] = [];
+  imageFiles: Blob[] = [];
+  @Output() uploadDone = new EventEmitter<boolean>();
 
-  ingredients: string[] = [];
-  instructions: string[] = [];
-  tools: string[] = [];
+  constructor(private fb: FormBuilder,
+              private db: AngularFirestore,
+              private router: Router,
+              private storage: AngularFireStorage,
+              private afs: AngularFirestore,
+              private _ngZone: NgZone,
+              private sanitizer: DomSanitizer,
+              private fAuth: AngularFireAuth) {}
 
-  recipeForm = new FormGroup({
-    name: new FormControl('', [Validators.required]),
-    description: new FormControl('', [Validators.required, Validators.maxLength(200)]),
-    tools: new FormControl(this.tools, [Validators.required]),
-    ingredients: new FormControl(this.ingredients, [Validators.required]),
-    instructions: new FormControl(this.instructions, [Validators.required]),
-    extraInfo: new FormControl(''),
-    difficulty: new FormControl('', [Validators.required]),
+  fileFormGroup = new FormGroup({
+    imageArray: new FormArray([], {validators: Validators.required} )
   });
 
-
-  addIngredient(newIngredient: string) {
-    var formattedIngredient: any = this.autoCapitalizeFirst(newIngredient);
-    if (formattedIngredient != '' && formattedIngredient != undefined) {
-      this.ingredients.push(formattedIngredient);
-    }
+  get imageArray() {
+    return this.fileFormGroup.get('imageArray') as FormArray;
   }
 
-  addInstruction(newInstruction: string) {
-    var onlyLettersStartInstruction: any =  this.removeNumsAndSymbolsAtStart(newInstruction);
-    if (onlyLettersStartInstruction != '') {
-      var formattedInstruction: any = this.autoCapitalizeFirst(onlyLettersStartInstruction);
-      if (formattedInstruction != '' && formattedInstruction != undefined) {
-        this.instructions.push(formattedInstruction);
-      } 
-    }
-  } 
-
-  addTool(newTool: string) {
-    var formattedTool: any = this.autoCapitalizeFirst(newTool);
-    if (formattedTool != '' && formattedTool != undefined) {
-      this.tools.push(formattedTool);
-    }
+  deleteImage(i: number) {
+    this.previewImgUrls.splice(i, 1);
+    this.imageArray.removeAt(i);
   }
 
-  autoCapitalizeFirst(fullString: string): string | undefined {
+  addImageToArray(event: any) {
+    if (event.target.files.length > 0) {
+      const file: Blob = event.target.files[0];
+      this.imgFile = file;
+      //uploads preview of image
+      //BypasssecuirtyTrustUrl creates a temporary URL for the known blob.
+      //This image/blob is one the user has just uploaded.
+      //It does not change the actual security of the URL, but allows the website to just display it temporarily. 
+      this.previewImgUrls.push(this.sanitizer.bypassSecurityTrustUrl(
+        URL.createObjectURL(file)
+      ));
+    }
+    this.imageArray.push(this.fb.control(event.target.value, [Validators.required, requiredFileType]));
+  }
+
+  basicsFormGroup = new FormGroup({
+    name: new FormControl('', Validators.required),
+    description: new FormControl('', Validators.required),
+    difficulty: new FormControl('', Validators.required),
+  });
+
+  extraFormGroup = new FormGroup({
+    extraInfo: new FormControl(''),
+  });
+
+  ingredientsFormGroup = new FormGroup({
+    ingredientsArray: new FormArray([this.fb.control('', Validators.required)], {validators: Validators.required}),
+  });
+
+  get ingredientsArray() {
+    return this.ingredientsFormGroup.get('ingredientsArray') as FormArray;
+  }
+  
+  addIngredientsField() {
+    this.ingredientsArray.push(this.fb.control('', Validators.required));
+  }
+
+  toolsFormGroup = new FormGroup({
+    toolsArray: new FormArray([this.fb.control('', Validators.required)]),
+  });
+  
+  get toolsArray() {
+    return this.toolsFormGroup.get('toolsArray') as FormArray;
+  }
+
+  addToolsField() {
+    this.toolsArray.push(this.fb.control('', Validators.required));
+  }
+
+  instructionsFormGroup = new FormGroup({
+    instructionsArray: new FormArray([this.fb.control('', Validators.required)]),
+  });
+  
+  get instructionsArray() {
+    return this.instructionsFormGroup.get('instructionsArray') as FormArray;
+  }
+
+  addInstructionsField() {
+    this.instructionsArray.push(this.fb.control('', Validators.required));
+  }
+
+
+
+  checkForBlanks(givenArray: string[]) {
+    for (let element of givenArray) {
+      if(element.trim() === '') {
+        return true;
+      }
+    }
+    return false;
+  }
+
+  checkInstructions(givenArray: string[]) {
+    for (let element of givenArray) {
+      if(this.removeNumsAndSymbolsAtStart(element) === '') {
+        return true;
+      }
+    }
+    return false;
+  }
+
+  autoCapitalizeFirst(fullString: string): string {
     var trimString: string = fullString.trim();
     if (trimString != '') {
       return trimString.charAt(0).toUpperCase() + trimString.substring(1);
     }
-    return undefined;
+    return '';
   }
 
   autoCapitalizeName(fullName: string): string {
@@ -76,44 +150,86 @@ export class UploadRecipeComponent {
     return formattedName;
   }
 
-  removeNumsAndSymbolsAtStart(originalString: string): string | undefined {
-      var startIndex: number = -1;
+  removeNumsAndSymbolsAtStart(originalString: string): string {
+    var startIndex: number = -1;
 
-      for (let i = 0; i < originalString.length; i++) {
-        var currentValue: number = originalString.charCodeAt(i);
-        if ((65 <= currentValue && currentValue <= 90) || (97 <= currentValue && currentValue <= 122)) {
-          return originalString.substring(i);
-        }
+    for (let i = 0; i < originalString.length; i++) {
+
+      // charCodeAt(i) of 65 is A and of 90 is Z
+      // charCodeA(i) of 97 is a and of 122 is z
+      var currentValue: number = originalString.charCodeAt(i);
+      if ((65 <= currentValue && currentValue <= 90) || (97 <= currentValue && currentValue <= 122)) {
+        return originalString.substring(i);
       }
-      if (startIndex == -1) {
-        return '';
-      }
+    }
+    if (startIndex == -1) {
       return '';
+    }
+    return '';
   }
 
-  onSubmit() {
-    if (this.autoCapitalizeFirst(this.recipeForm.value.extraInfo)) {
-      this.db.collection('recipes').add({
-        recipeName: this.autoCapitalizeName(this.recipeForm.value.name),
-        description: this.autoCapitalizeFirst(this.recipeForm.value.description),
-        ingredients: this.ingredients,
-        tools: this.tools,
-        instructions: this.instructions,
-        extraInfo: this.autoCapitalizeFirst(this.recipeForm.value.extraInfo),
-        difficulty: this.recipeForm.value.difficulty,
-        timestamp: Date.now(),
-      });
-    } else {
-      this.db.collection('recipes').add({
-        recipeName: this.autoCapitalizeName(this.recipeForm.value.name),
-        description: this.autoCapitalizeFirst(this.recipeForm.value.description),
-        ingredients: this.ingredients,
-        tools: this.tools,
-        instructions: this.instructions,
-        difficulty: this.recipeForm.value.difficulty,
-        timestamp: Date.now(),
-      });
+  formatArrays(originalArray: FormArray) {
+    for (let i = 0; i < originalArray.length; i++) {
+      originalArray.at(i).setValue(this.autoCapitalizeFirst(originalArray.at(i).value));
     }
+  }
+
+  formatInstructions(originalArray: FormArray) {
+    for (let i = 0; i < originalArray.length; i++) {
+      originalArray.at(i).setValue(this.autoCapitalizeFirst(this.removeNumsAndSymbolsAtStart(originalArray.at(i).value)));
+    }
+  }
+  
+  onSubmit() {
+    this.basicsFormGroup.controls.name.setValue(this.autoCapitalizeName(this.basicsFormGroup.value.name));
+    this.basicsFormGroup.controls.description.setValue(this.autoCapitalizeFirst(this.basicsFormGroup.value.description));
+    this.formatArrays(this.ingredientsArray);
+    this.formatArrays(this.toolsArray);
+    this.formatInstructions(this.instructionsArray);
+    this.extraFormGroup.controls.extraInfo.setValue(this.autoCapitalizeFirst(this.extraFormGroup.value.extraInfo));
+
+    this.db.collection('recipes')
+      .ref.withConverter(new RecipeConverter().recipeConverter)
+      .add(new Recipe(
+        this.basicsFormGroup.value.name,
+        '',
+        this.basicsFormGroup.value.difficulty,
+        this.basicsFormGroup.value.description,
+        this.ingredientsArray.value,
+        this.toolsArray.value,
+        [],
+        this.instructionsArray.value,
+        this.extraFormGroup.value.extraInfo ? this.extraFormGroup.value.extraInfo : '',
+        Date.now(),
+        [], 
+        [],
+    ));
     this.router.navigate(['/confirm-upload']);
   }
+
+  //image functions
+
+  onFileSubmit() {
+    this.fAuth.currentUser.then(user => {
+      if (user) {
+        //this.addImage(user.uid);
+      } else {
+        this._ngZone.run(() => {
+          this.router.navigate(['/home']);
+        });
+      }
+    });
+  }
+}
+
+function requiredFileType(control: FormControl) {
+  const imageTypes: Array<string> = ['png', 'jpg', 'jpeg', 'gif'];
+  const file = control.value;
+  if (file && file.split('.')[1]) {
+    const type = file.split('.')[1].toLowerCase();
+    if (imageTypes.indexOf(type) > -1) {
+      return null;
+    }
+  }
+  return {fileType: true};
 }
